+++ widgets/tws_widgets/lib/src/core/models/interfaces/tws_view_consume_adapter.dart
@@ -1,0 +1,7 @@
+
+import 'package:tws_foundation_client/tws_foundation_client.dart';
+
+/// [TWSViewConsumeAdapter] Consume adapter for sets view.
+abstract interface class TWSViewConsumeAdapter{
+  Future<List<SetViewOut<dynamic>>> consume(int page, int range, List<SetViewOrderOptions> orderings, String input);
+}