+++ widgets/tws_widgets/lib/src/core/models/models_module.dart
@@ -1,0 +1,4 @@
+library;
+
+export 'interfaces/tws_view_consume_adapter.dart';
+export 'tws_options_selector_action.dart';