+++ widgets/tws_widgets/lib/src/core/models/tws_options_selector_action.dart
@@ -1,0 +1,18 @@
+/// Custom class for action implementations on [TwsOptionSelector] widget.
+class TwsOptionSelectorAction<T>{
+  // Action title.
+  final String title;
+  // Action value. Returned when action is selected.
+  final T value;
+  // Min width.
+  final double minWidth;
+  // Max width
+  final double maxWidth;
+
+  const TwsOptionSelectorAction({
+    required this.title,
+    required this.value,
+    this.minWidth = double.maxFinite,
+    this.maxWidth = double.maxFinite,
+  });
+}