+++ widgets/tws_widgets/lib/src/themes/themes_module.dart
@@ -1,0 +1,6 @@
+library;
+
+// Themes exports -->
+export 'tws_foundation_dark_theme.dart';
+export 'tws_foundation_light_theme.dart';
+export 'tws_foundation_theme_base.dart';