+++ widgets/tws_widgets/lib/src/widgets/tws_section_divider.dart
@@ -1,0 +1,53 @@
+import 'package:csm_view/csm_view.dart';
+import 'package:flutter/material.dart';
+import 'package:tws_widgets/tws_widgets.dart';
+
+/// [TWSSectionDivider] a custom divider component to divide the sections or sub-sections.
+/// This component shows horizontal line with a centered section name, ideal for dividing sections in a column.
+class TWSSectionDivider extends StatelessWidget {
+  final Color? color;
+  final String text;
+  const TWSSectionDivider({
+    super.key,
+    this.color,
+    required this.text,
+  });
+
+  @override
+  Widget build(BuildContext context) {
+    Color? mainColor = color; 
+    if(color == null){
+      CSMColorThemeOptions pageColorTheme = getTheme<TWSFThemeBase>().page;
+      mainColor = pageColorTheme.fore;
+    }
+
+    return Padding(
+      padding: const EdgeInsets.all(5.0),
+      child: Row(
+        children: <Widget>[
+          Expanded(
+            child: Divider(
+              color: mainColor,
+            )
+          ),
+          Padding(
+            padding: const EdgeInsets.symmetric(horizontal: 8.0),
+            child: Text(
+              text,
+              style: TextStyle(
+                fontWeight: FontWeight.w100,
+                fontStyle: FontStyle.italic,
+                color: mainColor
+              ),
+            ),
+          ),
+          Expanded(
+            child: Divider(
+              color: mainColor,
+            )
+          )
+        ],
+      ),
+    );
+  }
+}