+++ widgets/tws_widgets/lib/src/widgets/widgets_module.dart
@@ -1,0 +1,22 @@
+library;
+
+// Exporting components -->
+export 'tws_button_flat.dart';
+export 'tws_cascade_section.dart';
+export 'tws_confirmation_dialog.dart';
+export 'tws_datepicker_field.dart';
+export 'tws_display_flat.dart';
+export 'tws_dropup.dart';
+export 'tws_file_picker.dart';
+export 'tws_frame_decoration.dart';
+export 'tws_image_viewer.dart';
+export 'tws_incremental_list.dart';
+export 'tws_input_text.dart';
+export 'tws_list_tile.dart';
+export 'tws_options_selector.dart';
+export 'tws_paging_selector.dart';
+export 'tws_property_viewer.dart';
+export 'tws_section.dart';
+export 'tws_section_divider.dart';
+export 'tws_selectable_list.dart';
+export 'tws_switch_button.dart';