+++ widgets/tws_widgets/lib/tws_widgets.dart
@@ -1,19 +1,5 @@
 library;
-
-// Export themes -->
-export 'src/themes/tws_foundation_dark_theme.dart';
-export 'src/themes/tws_foundation_light_theme.dart';
-export 'src/themes/tws_foundation_theme_base.dart';
-// Export components -->
-export 'src/widgets/tws_button_flat.dart';
-export 'src/widgets/tws_cascade_section.dart';
-export 'src/widgets/tws_confirmation_dialog.dart';
-export 'src/widgets/tws_datepicker_field.dart';
-export 'src/widgets/tws_display_flat.dart';
-export 'src/widgets/tws_dropup.dart';
-export 'src/widgets/tws_file_picker.dart';
-export 'src/widgets/tws_frame_decoration.dart';
-export 'src/widgets/tws_image_viewer.dart';
-export 'src/widgets/tws_incremental_list.dart';
-export 'src/widgets/tws_input_text.dart';
-export 'src/widgets/tws_section.dart';
+// Modules Export ---> 
+export 'src/core/models/models_module.dart';
+export 'src/themes/themes_module.dart';
+export 'src/widgets/widgets_module.dart';